+++ b2sdk/v1/file_version.py
@@ -1,13 +1,13 @@
 ######################################################################
 #
-# File: test/unit/v1/apiver/apiver_deps.py
+# File: b2sdk/v1/file_version.py
 #
-# Copyright 2020 Backblaze Inc. All Rights Reserved.
+# Copyright 2021 Backblaze Inc. All Rights Reserved.
 #
 # License https://www.backblaze.com/using_b2_code.html
 #
 ######################################################################
 
-from b2sdk.v1 import *  # noqa
+from b2sdk import _v2 as v2
 
-V = 1
+FileVersionInfo = v2.FileVersion