+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Refactored `minimum_part_size` to `recommended_part_size` (tha value used stays the same)
 * Refactored `sync.file.*File` and `sync.file.*FileVersion` to `sync.path.*SyncPath`
 * Encryption settings, types and providers are now part of the public API
+* Refactored `FileVersionInfo` to `FileVersion`
 
 ### Removed
 * Remove `Bucket.copy_file` and `Bucket.start_large_file`